refactor: decompose app-frame.ts into focused modules
Split the 1662-line app-frame.ts into 6 focused modules:

- app-frame.ts (526 lines) — public types + createFramedApp factory
- app-frame-types.ts (179 lines) — internal types, symbols, message utils
- app-frame-utils.ts (151 lines) — pure tree/layout/keymap utilities
- app-frame-render.ts (357 lines) — layout rendering + transitions
- app-frame-actions.ts (410 lines) — state reducers (tabs, scroll, panels)
- app-frame-palette.ts (200 lines) — command palette integration

Pure refactor: no public API changes, all 2017 tests pass unchanged.

Resolves bad_code.md item 6: app-frame.ts decomposition.
