+++ packages/bijou-tui/src/app-frame.ts
@@ -5,76 +5,68 @@
  * panel-scoped overlay context, and optional frame-level command palette.
  */
 
-import { resolveSafeCtx, type BijouContext } from '@flyingrobots/bijou';
-import { helpShort, helpView, type BindingSource } from './help.js';
-import {
-  createKeyMap,
-  formatKeyCombo,
-  type BindingInfo,
-  type KeyMap,
-} from './keybindings.js';
-import type { App, Cmd, KeyMsg } from './types.js';
-import { isKeyMsg, isMouseMsg, isResizeMsg, QUIT } from './types.js';
+import { resolveSafeCtx } from '@flyingrobots/bijou';
+import { helpView, type BindingSource } from './help.js';
+import type { KeyMap } from './keybindings.js';
+import type { App, Cmd } from './types.js';
+import { isKeyMsg, isMouseMsg, isResizeMsg } from './types.js';
 import type { Overlay } from './overlay.js';
 import { composite, modal } from './overlay.js';
 import type { TransitionShaderFn } from './transition-shaders.js';
 import { fitBlock } from './layout-utils.js';
-import { type BuiltinTransition, TRANSITION_SHADERS } from './transition-shaders.js';
+import { type BuiltinTransition } from './transition-shaders.js';
 import type { CommandPaletteItem, CommandPaletteState } from './command-palette.js';
 import {
   commandPalette,
   commandPaletteKeyMap,
-  createCommandPaletteState,
-  cpFilter,
-  cpFocusNext,
-  cpFocusPrev,
-  cpPageDown,
-  cpPageUp,
-  cpSelectedItem,
 } from './command-palette.js';
 import type { GridTrack } from './grid.js';
-import { grid, gridLayout } from './grid.js';
-import {
-  createFocusAreaState,
-  focusArea,
-  focusAreaPageDown,
-  focusAreaPageUp,
-  focusAreaScrollBy,
-  focusAreaScrollByX,
-  focusAreaScrollToBottom,
-  focusAreaScrollToTop,
-  focusAreaScrollTo,
-  focusAreaScrollToX,
-  type OverflowX,
-} from './focus-area.js';
-import { splitPane, splitPaneLayout, type SplitPaneDirection, type SplitPaneState } from './split-pane.js';
+import type { SplitPaneDirection, SplitPaneState } from './split-pane.js';
 import type { LayoutRect } from './layout-rect.js';
+import type { PanelVisibilityState } from './panel-state.js';
+import type { PanelMaximizeState } from './panel-state.js';
+import type { PanelDockState } from './panel-dock.js';
+import type { SerializedLayoutState } from './layout-preset.js';
+import { restoreLayoutState } from './layout-preset.js';
+import type { OverflowX } from './focus-area.js';
+import type { Timeline, TimelineState } from './timeline.js';
+
+// Internal modules
+import type {
+  InternalFrameModel,
+  PaletteAction,
+} from './app-frame-types.js';
 import {
-  createPanelVisibilityState,
-  createPanelMaximizeState,
-  toggleMinimized,
-  restorePane,
-  isMinimized,
-  toggleMaximize,
-  type PanelVisibilityState,
-  type PanelMaximizeState,
-} from './panel-state.js';
+  isFrameScopedMsg,
+  isPageScopedMsg,
+  wrapCmdForPage,
+  emitMsg,
+  emitMsgForPage,
+} from './app-frame-types.js';
 import {
-  createPanelDockState,
-  movePaneInContainer,
-  resolveChildOrder,
-  findPaneContainer,
-  getNodeId,
-  type PanelDockState,
-  type DockDirection,
-} from './panel-dock.js';
+  createFrameKeyMap,
+  frameBodyRect,
+  mergeBindingSources,
+} from './app-frame-utils.js';
 import {
-  restoreLayoutState,
-  type SerializedLayoutState,
-} from './layout-preset.js';
-import { clipToWidth, tokenizeAnsi, visibleLength } from './viewport.js';
-import { EASINGS } from './spring.js';
-import { timeline, type Timeline, type TimelineState } from './timeline.js';
+  renderHeaderLine,
+  renderHelpLine,
+  renderPageContent,
+  renderMaximizedPane,
+  renderTransition,
+} from './app-frame-render.js';
+import {
+  applyFrameAction,
+  syncPageFrameState,
+} from './app-frame-actions.js';
+import {
+  handlePaletteKey,
+  openCommandPalette,
+} from './app-frame-palette.js';
+
+// ---------------------------------------------------------------------------
+// Public types
+// ---------------------------------------------------------------------------
 
 /** Page declaration consumed by {@link createFramedApp}. */
 export interface FramePage<PageModel, Msg> {
@@ -250,86 +242,9 @@ export interface FrameModel<PageModel> {
   readonly splitRatioOverrides: Readonly<Record<string, Readonly<Record<string, number>>>>;
 }
 
-/** Internal model extending the public FrameModel with palette entries. */
-interface InternalFrameModel<PageModel, Msg> extends FrameModel<PageModel> {
-  readonly commandPaletteEntries?: readonly PaletteEntry<Msg>[];
-}
-
-/** A command palette entry linking a UI item to an action or message. */
-interface PaletteEntry<Msg> {
-  readonly id: string;
-  readonly item: CommandPaletteItem;
-  readonly msgAction?: Msg;
-  readonly targetPageId?: string;
-  readonly frameAction?: FrameAction;
-}
-
-/** Per-call context passed through the recursive layout renderer. */
-interface RenderContext<PageModel, Msg> {
-  readonly model: InternalFrameModel<PageModel, Msg>;
-  readonly pageId: string;
-  readonly focusedPaneId: string | undefined;
-  readonly scrollByPane: Readonly<Record<string, FramePaneScroll>>;
-  readonly visibility: PanelVisibilityState;
-  readonly dockState: PanelDockState;
-}
-
-/** Output of a layout node render pass. */
-interface RenderResult {
-  readonly output: string;
-  readonly paneRects: ReadonlyMap<string, LayoutRect>;
-  readonly paneOrder: readonly string[];
-}
-
-/** Discriminated union of all frame-level actions (tabs, panes, scroll, palette, help, transitions). */
-type FrameAction =
-  | { type: 'toggle-help' }
-  | { type: 'prev-tab' }
-  | { type: 'next-tab' }
-  | { type: 'next-pane' }
-  | { type: 'prev-pane' }
-  | { type: 'scroll-up' }
-  | { type: 'scroll-down' }
-  | { type: 'page-up' }
-  | { type: 'page-down' }
-  | { type: 'top' }
-  | { type: 'bottom' }
-  | { type: 'scroll-left' }
-  | { type: 'scroll-right' }
-  | { type: 'open-palette' }
-  | { type: 'toggle-minimize' }
-  | { type: 'toggle-maximize' }
-  | { type: 'dock-up' }
-  | { type: 'dock-down' }
-  | { type: 'dock-left' }
-  | { type: 'dock-right' }
-  | { type: 'transition'; progress: number; generation: number }
-  | { type: 'transition-complete'; generation: number };
-
-/** Discriminated union of command palette navigation/selection actions. */
-type PaletteAction =
-  | { type: 'cp-next' }
-  | { type: 'cp-prev' }
-  | { type: 'cp-page-down' }
-  | { type: 'cp-page-up' }
-  | { type: 'cp-select' }
-  | { type: 'cp-close' };
-
-const PAGE_MSG_TOKEN = Symbol('app-frame-page-msg');
-const FRAME_MSG_TOKEN = Symbol('app-frame-frame-msg');
-
-/** Wrapper that tags a user message with its originating page ID. */
-interface PageScopedMsg<Msg> {
-  readonly [PAGE_MSG_TOKEN]: true;
-  readonly pageId: string;
-  readonly msg: Msg;
-}
-
-/** Wrapper that tags a frame-internal action for the update loop. */
-interface FrameScopedMsg {
-  readonly [FRAME_MSG_TOKEN]: true;
-  readonly action: FrameAction;
-}
+// ---------------------------------------------------------------------------
+// Factory
+// ---------------------------------------------------------------------------
 
 /**
  * Create a fully framed TEA app shell.
@@ -488,6 +403,10 @@ export function createFramedApp<PageModel, Msg>(
 
         const frameAction = frameKeys.handle(msg);
         if (frameAction !== undefined) {
+          // Handle palette opening here since applyFrameAction doesn't have access to palette deps
+          if (frameAction.type === 'open-palette' && options.enableCommandPalette) {
+            return [openCommandPalette(model, frameKeys, options, pagesById), []];
+          }
           return applyFrameAction(frameAction, model, frameKeys, options, pagesById);
         }
 
@@ -604,1059 +523,4 @@ export function createFramedApp<PageModel, Msg>(
   };
 
   return app;
-}
-
-/** Route a key press through the command palette, returning updated model and commands. */
-function handlePaletteKey<PageModel, Msg>(
-  msg: KeyMsg,
-  model: InternalFrameModel<PageModel, Msg>,
-  paletteKeys: KeyMap<PaletteAction>,
-  frameKeys: KeyMap<FrameAction>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): [InternalFrameModel<PageModel, Msg>, Cmd<Msg>[]] {
-  const cp = model.commandPalette!;
-  const action = paletteKeys.handle(msg);
-
-  if (action != null) {
-    switch (action.type) {
-      case 'cp-next':
-        return [{ ...model, commandPalette: cpFocusNext(cp) }, []];
-      case 'cp-prev':
-        return [{ ...model, commandPalette: cpFocusPrev(cp) }, []];
-      case 'cp-page-down':
-        return [{ ...model, commandPalette: cpPageDown(cp) }, []];
-      case 'cp-page-up':
-        return [{ ...model, commandPalette: cpPageUp(cp) }, []];
-      case 'cp-close':
-        return [{ ...model, commandPalette: undefined, commandPaletteEntries: undefined }, []];
-      case 'cp-select': {
-        const selected = cpSelectedItem(cp);
-        if (selected == null) {
-          return [{ ...model, commandPalette: undefined, commandPaletteEntries: undefined }, []];
-        }
-        const entry = model.commandPaletteEntries?.find((x) => x.id === selected.id);
-        if (entry?.frameAction != null) {
-          const closed = { ...model, commandPalette: undefined, commandPaletteEntries: undefined };
-          return applyFrameAction(entry.frameAction, closed, frameKeys, options, pagesById);
-        }
-        if (entry?.msgAction !== undefined) {
-          const cmd = entry.targetPageId != null
-            ? emitMsgForPage(entry.targetPageId, entry.msgAction)
-            : emitMsg(entry.msgAction);
-          return [{
-            ...model,
-            commandPalette: undefined,
-            commandPaletteEntries: undefined,
-          }, [cmd]];
-        }
-        return [{ ...model, commandPalette: undefined, commandPaletteEntries: undefined }, []];
-      }
-    }
-  }
-
-  if (msg.key === 'backspace') {
-    const next = cpFilter(cp, cp.query.slice(0, -1));
-    return [{ ...model, commandPalette: next }, []];
-  }
-
-  if (msg.ctrl && msg.key === 'c') {
-    return [{ ...model, commandPalette: undefined, commandPaletteEntries: undefined }, []];
-  }
-
-  if (!msg.ctrl && !msg.alt && !msg.shift && msg.key === 'q' && cp.query.length === 0) {
-    return [{ ...model, commandPalette: undefined, commandPaletteEntries: undefined }, []];
-  }
-
-  if (!msg.ctrl && !msg.alt && msg.key.length === 1) {
-    const next = cpFilter(cp, cp.query + msg.key);
-    return [{ ...model, commandPalette: next }, []];
-  }
-
-  return [model, []];
-}
-
-/** Dispatch a frame-level action (tab switch, pane cycle, scroll, palette, help toggle, transitions). */
-function applyFrameAction<PageModel, Msg>(
-  action: FrameAction,
-  model: InternalFrameModel<PageModel, Msg>,
-  frameKeys: KeyMap<FrameAction>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): [InternalFrameModel<PageModel, Msg>, Cmd<Msg>[]] {
-  switch (action.type) {
-    case 'toggle-help':
-      return [{ ...model, helpOpen: !model.helpOpen }, []];
-    case 'prev-tab':
-      return switchTab(model, -1, pagesById, options);
-    case 'next-tab':
-      return switchTab(model, 1, pagesById, options);
-    case 'next-pane':
-      return [cyclePane(model, 1, pagesById), []];
-    case 'prev-pane':
-      return [cyclePane(model, -1, pagesById), []];
-    case 'open-palette':
-      if (!options.enableCommandPalette) return [model, []];
-      return [openCommandPalette(model, frameKeys, options, pagesById), []];
-    case 'toggle-minimize':
-      return [applyToggleMinimize(model, pagesById), []];
-    case 'toggle-maximize':
-      return [applyToggleMaximize(model, pagesById), []];
-    case 'dock-up':
-    case 'dock-down':
-    case 'dock-left':
-    case 'dock-right': {
-      const dir = action.type.replace('dock-', '') as DockDirection;
-      return [applyDockMove(model, dir, pagesById), []];
-    }
-    case 'scroll-up':
-    case 'scroll-down':
-    case 'page-up':
-    case 'page-down':
-    case 'top':
-    case 'bottom':
-    case 'scroll-left':
-    case 'scroll-right':
-      return [scrollFocusedPane(model, action, pagesById), []];
-    case 'transition':
-    case 'transition-complete':
-      return [model, []];
-  }
-}
-
-/** Cycle the active tab by `delta` positions, optionally starting a transition. */
-function switchTab<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  delta: number,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-): [InternalFrameModel<PageModel, Msg>, Cmd<Msg>[]] {
-  const idx = model.pageOrder.indexOf(model.activePageId);
-  if (idx < 0) return [model, []];
-  const nextIdx = (idx + delta + model.pageOrder.length) % model.pageOrder.length;
-  const nextId = model.pageOrder[nextIdx]!;
-
-  if (nextId === model.activePageId) return [model, []];
-
-  const activePageModel = model.pageModels[model.activePageId]!;
-  const activeTransition = options.transitionOverride
-    ? options.transitionOverride(activePageModel)
-    : options.transition;
-
-  const hasTransition = activeTransition != null && activeTransition !== 'none';
-  const nextGeneration = model.transitionGeneration + 1;
-
-  // Use the user-supplied timeline if provided, otherwise build a default.
-  // The timeline MUST contain a 'progress' track (0 → 1).
-  const tl = hasTransition
-    ? (options.transitionTimeline ?? timeline()
-      .add('progress', {
-        type: 'tween',
-        from: 0,
-        to: 1,
-        duration: options.transitionDuration ?? 300,
-        ease: EASINGS.easeInOutCubic,
-      })
-      .build())
-    : undefined;
-
-  const durationMs = tl?.estimatedDurationMs ?? options.transitionDuration ?? 300;
-
-  const nextModel = syncPageFrameState({
-    ...model,
-    activePageId: nextId,
-    previousPageId: model.activePageId,
-    activeTransition,
-    transitionProgress: hasTransition ? 0 : 1,
-    transitionGeneration: nextGeneration,
-    transitionStartMs: hasTransition ? Date.now() : undefined,
-    transitionTimeline: tl,
-    transitionTimelineState: tl?.init(),
-  }, nextId, pagesById);
-
-  if (hasTransition) {
-    // Schedule render ticks at ~60fps for the duration of the transition.
-    // Each tick advances the timeline using wall-clock elapsed time.
-    const cmd: Cmd<Msg> = createTransitionTickCmd(durationMs, nextGeneration);
-    return [nextModel, [cmd]];
-  }
-
-  return [nextModel, []];
-}
-
-/** Move focus to the next or previous pane in the active page's layout. */
-function cyclePane<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  delta: number,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const page = pagesById.get(model.activePageId)!;
-  const paneIds = collectPaneIds(page.layout(model.pageModels[model.activePageId]!));
-  if (paneIds.length === 0) return model;
-
-  const curr = model.focusedPaneByPage[model.activePageId];
-  const idx = curr == null ? 0 : paneIds.indexOf(curr);
-  const nextIdx = idx < 0
-    ? 0
-    : (idx + delta + paneIds.length) % paneIds.length;
-  const next = paneIds[nextIdx]!;
-  return {
-    ...model,
-    focusedPaneByPage: {
-      ...model.focusedPaneByPage,
-      [model.activePageId]: next,
-    },
-  };
-}
-
-/** Apply a scroll action to the currently focused pane. */
-function scrollFocusedPane<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  action: Extract<
-    FrameAction,
-    { type: 'scroll-up' | 'scroll-down' | 'page-up' | 'page-down' | 'top' | 'bottom' | 'scroll-left' | 'scroll-right' }
-  >,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const pageId = model.activePageId;
-  const focusedPaneId = model.focusedPaneByPage[pageId];
-  if (focusedPaneId == null) return model;
-
-  const page = pagesById.get(pageId)!;
-  const layoutTree = page.layout(model.pageModels[pageId]!);
-  const bodyRect = frameBodyRect(model.columns, model.rows);
-  const resolved = renderFrameNode(layoutTree, bodyRect, {
-    model,
-    pageId,
-    focusedPaneId,
-    scrollByPane: model.scrollByPage[pageId] ?? {},
-    visibility: model.minimizedByPage[pageId] ?? createPanelVisibilityState(),
-    dockState: model.dockStateByPage[pageId] ?? createPanelDockState(),
-  });
-  const paneRect = resolved.paneRects.get(focusedPaneId);
-  if (paneRect == null || paneRect.width <= 0 || paneRect.height <= 0) return model;
-
-  const paneNode = findPaneNode(layoutTree, focusedPaneId);
-  if (paneNode == null) return model;
-
-  const content = paneNode.render(paneRect.width, paneRect.height);
-  let state = createFocusAreaState({
-    content,
-    width: paneRect.width,
-    height: paneRect.height,
-    overflowX: paneNode.overflowX ?? 'hidden',
-  });
-  const prior = model.scrollByPage[pageId]?.[focusedPaneId] ?? { x: 0, y: 0 };
-  state = focusAreaScrollTo(state, prior.y);
-  state = focusAreaScrollToX(state, prior.x);
-
-  switch (action.type) {
-    case 'scroll-up':
-      state = focusAreaScrollBy(state, -1);
-      break;
-    case 'scroll-down':
-      state = focusAreaScrollBy(state, 1);
-      break;
-    case 'page-up':
-      state = focusAreaPageUp(state);
-      break;
-    case 'page-down':
-      state = focusAreaPageDown(state);
-      break;
-    case 'top':
-      state = focusAreaScrollToTop(state);
-      break;
-    case 'bottom':
-      state = focusAreaScrollToBottom(state);
-      break;
-    case 'scroll-left':
-      state = focusAreaScrollByX(state, -1);
-      break;
-    case 'scroll-right':
-      state = focusAreaScrollByX(state, 1);
-      break;
-  }
-
-  const pageScroll = model.scrollByPage[pageId] ?? {};
-  return {
-    ...model,
-    scrollByPage: {
-      ...model.scrollByPage,
-      [pageId]: {
-        ...pageScroll,
-        [focusedPaneId]: { x: state.scroll.x, y: state.scroll.y },
-      },
-    },
-  };
-}
-
-/** Initialize the command palette with entries from frame, global, page key maps, and custom page command items. */
-function openCommandPalette<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  frameKeys: KeyMap<FrameAction>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const entries = buildPaletteEntries(model, frameKeys, options, pagesById);
-  const items = entries.map((x) => x.item);
-  return {
-    ...model,
-    commandPalette: createCommandPaletteState(items, Math.max(5, Math.min(10, model.rows - 8))),
-    commandPaletteEntries: entries,
-  };
-}
-
-/** Collect all available commands from frame, global, page, and custom sources. */
-function buildPaletteEntries<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  frameKeys: KeyMap<FrameAction>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): readonly PaletteEntry<Msg>[] {
-  const entries: PaletteEntry<Msg>[] = [];
-  let seq = 0;
-
-  for (const b of frameKeys.bindings()) {
-    if (!b.enabled) continue;
-    const action = frameKeys.handle(comboToMsg(b));
-    if (action === undefined) continue;
-    const id = `frame:${seq++}`;
-    entries.push({
-      id,
-      item: {
-        id,
-        label: b.description,
-        category: 'Frame',
-        shortcut: formatKeyCombo(b.combo),
-      },
-      frameAction: action,
-    });
-  }
-
-  const global = options.globalKeys;
-  if (global != null) {
-    for (const b of global.bindings()) {
-      if (!b.enabled) continue;
-      const action = global.handle(comboToMsg(b));
-      if (action === undefined) continue;
-      const id = `global:${seq++}`;
-      entries.push({
-        id,
-        item: {
-          id,
-          label: b.description,
-          category: 'Global',
-          shortcut: formatKeyCombo(b.combo),
-        },
-        msgAction: action,
-      });
-    }
-  }
-
-  const page = pagesById.get(model.activePageId)!;
-  if (page.keyMap != null) {
-    for (const b of page.keyMap.bindings()) {
-      if (!b.enabled) continue;
-      const action = page.keyMap.handle(comboToMsg(b));
-      if (action === undefined) continue;
-      const id = `page:${seq++}`;
-      entries.push({
-        id,
-        item: {
-          id,
-          label: b.description,
-          category: page.title,
-          shortcut: formatKeyCombo(b.combo),
-        },
-        msgAction: action,
-        targetPageId: model.activePageId,
-      });
-    }
-  }
-
-  if (page.commandItems != null) {
-    for (const item of page.commandItems(model.pageModels[model.activePageId]!)) {
-      const id = `custom:${seq++}`;
-      entries.push({
-        id,
-        item: {
-          ...item,
-          id,
-          category: item.category ?? page.title,
-        },
-        msgAction: item.action,
-        targetPageId: model.activePageId,
-      });
-    }
-  }
-
-  return entries;
-}
-
-/** Reconcile pane IDs, scroll offsets, and focus for a page after init, tab switches, or window resizes. */
-function syncPageFrameState<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  pageId: string,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const page = pagesById.get(pageId)!;
-  const paneIds = collectPaneIds(page.layout(model.pageModels[pageId]!));
-  assertUniquePaneIds(paneIds, `page "${pageId}" layout`);
-
-  const prevScroll = model.scrollByPage[pageId] ?? {};
-  const nextScroll: Record<string, FramePaneScroll> = {};
-  for (const paneId of paneIds) {
-    nextScroll[paneId] = prevScroll[paneId] ?? { x: 0, y: 0 };
-  }
-
-  const prevFocused = model.focusedPaneByPage[pageId];
-  const focused = prevFocused != null && paneIds.includes(prevFocused)
-    ? prevFocused
-    : paneIds[0];
-
-  return {
-    ...model,
-    focusedPaneByPage: {
-      ...model.focusedPaneByPage,
-      [pageId]: focused,
-    },
-    scrollByPage: {
-      ...model.scrollByPage,
-      [pageId]: nextScroll,
-    },
-  };
-}
-
-/** Recursively render a layout tree node (pane, split, or grid) into a rect. */
-function renderFrameNode<PageModel, Msg>(
-  node: FrameLayoutNode,
-  rect: LayoutRect,
-  ctx: RenderContext<PageModel, Msg>,
-): RenderResult {
-  if (rect.width <= 0 || rect.height <= 0) {
-    return { output: '', paneRects: new Map(), paneOrder: [] };
-  }
-
-  if (node.kind === 'pane') {
-    // Minimized pane: render as collapsed title bar
-    if (isMinimized(ctx.visibility, node.paneId)) {
-      const titleBar = `[${node.paneId}] \u25b8`; // ▸
-      const output = fitBlock(titleBar, rect.width, rect.height).join('\n');
-      return {
-        output,
-        paneRects: new Map([[node.paneId, rect]]),
-        paneOrder: [node.paneId],
-      };
-    }
-
-    const prior = ctx.scrollByPane[node.paneId] ?? { x: 0, y: 0 };
-    const content = node.render(rect.width, rect.height);
-    let state = createFocusAreaState({
-      content,
-      width: rect.width,
-      height: rect.height,
-      overflowX: node.overflowX ?? 'hidden',
-    });
-    state = focusAreaScrollTo(state, prior.y);
-    state = focusAreaScrollToX(state, prior.x);
-    const output = focusArea(state, { focused: node.paneId === ctx.focusedPaneId });
-    return {
-      output,
-      paneRects: new Map([[node.paneId, rect]]),
-      paneOrder: [node.paneId],
-    };
-  }
-
-  if (node.kind === 'split') {
-    const direction = node.direction ?? 'row';
-
-    // Consult dock state for child order
-    const defaultChildIds = [getNodeId(node.paneA), getNodeId(node.paneB)];
-    const resolvedOrder = resolveChildOrder(ctx.dockState, node.splitId, defaultChildIds);
-    const swapped = resolvedOrder[0] !== defaultChildIds[0];
-    const effectiveA = swapped ? node.paneB : node.paneA;
-    const effectiveB = swapped ? node.paneA : node.paneB;
-
-    // Check for minimized panes — give minimized pane 1 row, sibling gets rest
-    const aMinimized = isPaneMinimized(effectiveA, ctx.visibility);
-    const bMinimized = isPaneMinimized(effectiveB, ctx.visibility);
-
-    let splitState = node.state;
-    // Apply split ratio overrides from layout presets
-    const ratioOverride = ctx.model.splitRatioOverrides[ctx.pageId]?.[node.splitId];
-    if (ratioOverride !== undefined) {
-      splitState = { ...splitState, ratio: ratioOverride };
-    }
-
-    if (aMinimized && !bMinimized) {
-      // A is minimized: give it minimal space
-      const mainAxisTotal = direction === 'row' ? rect.width : rect.height;
-      const minimizedSize = Math.min(1, mainAxisTotal);
-      splitState = { ...splitState, ratio: minimizedSize / Math.max(1, mainAxisTotal) };
-    } else if (bMinimized && !aMinimized) {
-      // B is minimized: give A most of the space
-      const mainAxisTotal = direction === 'row' ? rect.width : rect.height;
-      const minimizedSize = Math.min(1, mainAxisTotal);
-      splitState = { ...splitState, ratio: 1 - minimizedSize / Math.max(1, mainAxisTotal) };
-    }
-
-    const layout = splitPaneLayout(splitState, {
-      direction,
-      width: rect.width,
-      height: rect.height,
-      minA: node.minA,
-      minB: node.minB,
-    });
-
-    const aRect = offsetRect(layout.paneA, rect.row, rect.col);
-    const bRect = offsetRect(layout.paneB, rect.row, rect.col);
-
-    const a = renderFrameNode(effectiveA, aRect, ctx);
-    const b = renderFrameNode(effectiveB, bRect, ctx);
-
-    const output = splitPane(splitState, {
-      direction,
-      width: rect.width,
-      height: rect.height,
-      minA: node.minA,
-      minB: node.minB,
-      dividerChar: node.dividerChar,
-      paneA: () => a.output,
-      paneB: () => b.output,
-    });
-
-    return {
-      output,
-      paneRects: mergeMaps(a.paneRects, b.paneRects),
-      paneOrder: [...a.paneOrder, ...b.paneOrder],
-    };
-  }
-
-  const relRects = gridLayout({
-    width: rect.width,
-    height: rect.height,
-    columns: node.columns,
-    rows: node.rows,
-    areas: node.areas,
-    gap: node.gap,
-  });
-
-  const renderedByArea = new Map<string, RenderResult>();
-  for (const [areaName, areaRect] of relRects) {
-    const absoluteAreaRect = offsetRect(areaRect, rect.row, rect.col);
-    const child = node.cells[areaName];
-    if (child == null) {
-      resolveSafeCtx()?.io.writeError(
-        `createFramedApp: grid cell "${areaName}" missing in page "${ctx.pageId}" — rendering placeholder\n`,
-      );
-      renderedByArea.set(areaName, renderMissingGridCell(areaName, absoluteAreaRect));
-      continue;
-    }
-    renderedByArea.set(areaName, renderFrameNode(child, absoluteAreaRect, ctx));
-  }
-
-  const output = grid({
-    width: rect.width,
-    height: rect.height,
-    columns: node.columns,
-    rows: node.rows,
-    areas: node.areas,
-    gap: node.gap,
-    cells: Object.fromEntries([...renderedByArea.entries()].map(([name, rendered]) => [name, () => rendered.output])),
-  });
-
-  let paneRects = new Map<string, LayoutRect>();
-  const seenPaneIds = new Set<string>();
-  const paneOrder: string[] = [];
-  for (const rendered of renderedByArea.values()) {
-    for (const [paneId, paneRect] of rendered.paneRects.entries()) {
-      if (paneRects.has(paneId)) {
-        throw new Error(`createFramedApp: duplicate paneId "${paneId}" in rendered layout`);
-      }
-      paneRects.set(paneId, paneRect);
-    }
-    for (const paneId of rendered.paneOrder) {
-      if (seenPaneIds.has(paneId)) {
-        throw new Error(`createFramedApp: duplicate paneId "${paneId}" in rendered pane order`);
-      }
-      seenPaneIds.add(paneId);
-      paneOrder.push(paneId);
-    }
-  }
-
-  return { output, paneRects, paneOrder };
-}
-
-/** Render a placeholder for a grid area that has no matching cell definition. */
-function renderMissingGridCell(areaName: string, rect: LayoutRect): RenderResult {
-  return {
-    output: fitBlock(`[missing grid cell: ${areaName}]`, rect.width, rect.height).join('\n'),
-    paneRects: new Map(),
-    paneOrder: [],
-  };
-}
-
-/** Recursively collect all pane IDs from a layout tree in declaration order. */
-function collectPaneIds(node: FrameLayoutNode): string[] {
-  if (node.kind === 'pane') return [node.paneId];
-  if (node.kind === 'split') return [...collectPaneIds(node.paneA), ...collectPaneIds(node.paneB)];
-
-  const ids: string[] = [];
-  for (const areaName of declaredAreaNames(node.areas)) {
-    const child = node.cells[areaName];
-    if (child == null) continue;
-    ids.push(...collectPaneIds(child));
-  }
-  return ids;
-}
-
-/** Extract unique area names from CSS-grid-style template strings. */
-function declaredAreaNames(areas: readonly string[]): string[] {
-  const names = new Set<string>();
-  for (const row of areas) {
-    for (const token of row.trim().split(/\s+/)) {
-      if (token !== '' && token !== '.') names.add(token);
-    }
-  }
-  return [...names];
-}
-
-/** Throw if any pane ID appears more than once in the given list. */
-function assertUniquePaneIds(paneIds: readonly string[], scope: string): void {
-  const seen = new Set<string>();
-  for (const paneId of paneIds) {
-    if (seen.has(paneId)) {
-      throw new Error(`createFramedApp: duplicate paneId "${paneId}" in ${scope}`);
-    }
-    seen.add(paneId);
-  }
-}
-
-/** Toggle minimize on the focused pane of the active page. */
-function applyToggleMinimize<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const pageId = model.activePageId;
-  const focusedPaneId = model.focusedPaneByPage[pageId];
-  if (focusedPaneId == null) return model;
-
-  const page = pagesById.get(pageId)!;
-  const allPaneIds = collectPaneIds(page.layout(model.pageModels[pageId]!));
-  const current = model.minimizedByPage[pageId] ?? createPanelVisibilityState();
-  const next = toggleMinimized(current, focusedPaneId, allPaneIds);
-
-  // If we just minimized the focused pane, move focus to a non-minimized sibling
-  let newFocused = focusedPaneId;
-  if (isMinimized(next, focusedPaneId)) {
-    const visible = allPaneIds.filter((id) => !isMinimized(next, id));
-    newFocused = visible[0] ?? focusedPaneId;
-  }
-
-  return {
-    ...model,
-    minimizedByPage: { ...model.minimizedByPage, [pageId]: next },
-    focusedPaneByPage: { ...model.focusedPaneByPage, [pageId]: newFocused },
-  };
-}
-
-/** Toggle maximize on the focused pane of the active page. */
-function applyToggleMaximize<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  _pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const pageId = model.activePageId;
-  const focusedPaneId = model.focusedPaneByPage[pageId];
-  if (focusedPaneId == null) return model;
-
-  const current = model.maximizedPaneByPage[pageId] ?? createPanelMaximizeState();
-  const next = toggleMaximize(current, focusedPaneId);
-
-  // If maximizing a minimized pane, restore it first
-  let visibility = model.minimizedByPage[pageId] ?? createPanelVisibilityState();
-  if (next.maximizedPaneId && isMinimized(visibility, next.maximizedPaneId)) {
-    visibility = restorePane(visibility, next.maximizedPaneId);
-  }
-
-  return {
-    ...model,
-    maximizedPaneByPage: { ...model.maximizedPaneByPage, [pageId]: next },
-    minimizedByPage: { ...model.minimizedByPage, [pageId]: visibility },
-  };
-}
-
-/** Move the focused pane within its container in the given direction. */
-function applyDockMove<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  direction: DockDirection,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): InternalFrameModel<PageModel, Msg> {
-  const pageId = model.activePageId;
-  const focusedPaneId = model.focusedPaneByPage[pageId];
-  if (focusedPaneId == null) return model;
-
-  const page = pagesById.get(pageId)!;
-  const layoutTree = page.layout(model.pageModels[pageId]!);
-  const container = findPaneContainer(layoutTree, focusedPaneId);
-  if (container == null) return model;
-
-  const dockState = model.dockStateByPage[pageId] ?? createPanelDockState();
-  const currentOrder = resolveChildOrder(dockState, container.containerId, container.childIds);
-  const next = movePaneInContainer(dockState, container.containerId, focusedPaneId, direction, currentOrder);
-
-  return {
-    ...model,
-    dockStateByPage: { ...model.dockStateByPage, [pageId]: next },
-  };
-}
-
-/** Render only the maximized pane at the full body rect. */
-function renderMaximizedPane<PageModel, Msg>(
-  pageId: string,
-  model: InternalFrameModel<PageModel, Msg>,
-  bodyRect: LayoutRect,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-  maximizedPaneId: string,
-): RenderResult {
-  const page = pagesById.get(pageId)!;
-  const pageModel = model.pageModels[pageId]!;
-  const layoutTree = page.layout(pageModel);
-  const paneNode = findPaneNode(layoutTree, maximizedPaneId);
-  if (paneNode == null) {
-    // Pane not found, fall back to normal rendering
-    return renderPageContent(pageId, model, bodyRect, pagesById);
-  }
-
-  const prior = model.scrollByPage[pageId]?.[maximizedPaneId] ?? { x: 0, y: 0 };
-  const content = paneNode.render(bodyRect.width, bodyRect.height);
-  let state = createFocusAreaState({
-    content,
-    width: bodyRect.width,
-    height: bodyRect.height,
-    overflowX: paneNode.overflowX ?? 'hidden',
-  });
-  state = focusAreaScrollTo(state, prior.y);
-  state = focusAreaScrollToX(state, prior.x);
-  const output = focusArea(state, { focused: true });
-
-  return {
-    output,
-    paneRects: new Map([[maximizedPaneId, bodyRect]]),
-    paneOrder: [maximizedPaneId],
-  };
-}
-
-/** Walk the layout tree to find the pane node with the given ID. */
-function findPaneNode(node: FrameLayoutNode, paneId: string): Extract<FrameLayoutNode, { kind: 'pane' }> | undefined {
-  if (node.kind === 'pane') return node.paneId === paneId ? node : undefined;
-  if (node.kind === 'split') return findPaneNode(node.paneA, paneId) ?? findPaneNode(node.paneB, paneId);
-  for (const key of Object.keys(node.cells)) {
-    const found = findPaneNode(node.cells[key]!, paneId);
-    if (found) return found;
-  }
-  return undefined;
-}
-
-/** Build the default key map for frame-level actions (tabs, panes, scroll, help '?', command palette 'ctrl+p'/':'). */
-function createFrameKeyMap(): KeyMap<FrameAction> {
-  return createKeyMap<FrameAction>()
-    .group('Frame', (g) => g
-      .bind('?', 'Toggle help', { type: 'toggle-help' })
-      .bind('[', 'Previous tab', { type: 'prev-tab' })
-      .bind(']', 'Next tab', { type: 'next-tab' })
-      .bind('tab', 'Next pane', { type: 'next-pane' })
-      .bind('shift+tab', 'Previous pane', { type: 'prev-pane' })
-      .bind('ctrl+p', 'Open command palette', { type: 'open-palette' })
-      .bind(':', 'Open command palette', { type: 'open-palette' })
-      .bind('ctrl+m', 'Fold/unfold pane', { type: 'toggle-minimize' })
-      .bind('ctrl+f', 'Full-screen pane', { type: 'toggle-maximize' }),
-    )
-    .group('Dock', (g) => g
-      .bind('ctrl+shift+up', 'Move pane up', { type: 'dock-up' })
-      .bind('ctrl+shift+down', 'Move pane down', { type: 'dock-down' })
-      .bind('ctrl+shift+left', 'Move pane left', { type: 'dock-left' })
-      .bind('ctrl+shift+right', 'Move pane right', { type: 'dock-right' }),
-    )
-    .group('Scroll', (g) => g
-      .bind('j', 'Scroll down', { type: 'scroll-down' })
-      .bind('k', 'Scroll up', { type: 'scroll-up' })
-      .bind('d', 'Page down', { type: 'page-down' })
-      .bind('u', 'Page up', { type: 'page-up' })
-      .bind('g', 'Top', { type: 'top' })
-      .bind('shift+g', 'Bottom', { type: 'bottom' })
-      .bind('h', 'Scroll left', { type: 'scroll-left' })
-      .bind('l', 'Scroll right', { type: 'scroll-right' }),
-    );
-}
-
-/** Render the top header line showing the app title and tab bar. */
-function renderHeaderLine<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): string {
-  const title = options.title ?? 'App';
-  const tabs = model.pageOrder.map((id) => {
-    const page = pagesById.get(id)!;
-    return id === model.activePageId ? `[${page.title}]` : ` ${page.title} `;
-  }).join(' ');
-
-  return fitLine(`${title}  ${tabs}`, model.columns);
-}
-
-/** Render the bottom status line showing mode, focused pane, and key hints. */
-function renderHelpLine<PageModel, Msg>(
-  model: InternalFrameModel<PageModel, Msg>,
-  frameKeys: KeyMap<FrameAction>,
-  options: CreateFramedAppOptions<PageModel, Msg>,
-  activePage: FramePage<PageModel, Msg>,
-): string {
-  const mode = model.commandPalette != null
-    ? 'PALETTE'
-    : model.helpOpen
-      ? 'HELP'
-      : 'NORMAL';
-  const focusedPane = model.focusedPaneByPage[model.activePageId] ?? '-';
-  const status = `[${mode}] page:${model.activePageId} pane:${focusedPane}`;
-
-  const source = mergeBindingSources(
-    frameKeys,
-    options.globalKeys,
-    activePage.helpSource ?? activePage.keyMap,
-  );
-  const hint = helpShort(source);
-  const line = hint.length > 0
-    ? ` ${status}  ${hint}`
-    : ` ${status}`;
-  return fitLine(line, model.columns);
-}
-
-/** Combine multiple binding sources into a single source for help display. */
-function mergeBindingSources(...sources: Array<BindingSource | undefined>): BindingSource {
-  return {
-    bindings(): readonly BindingInfo[] {
-      const merged: BindingInfo[] = [];
-      for (const src of sources) {
-        if (src == null) continue;
-        merged.push(...src.bindings());
-      }
-      return merged;
-    },
-  };
-}
-
-/** Compute the available rect for page content (screen minus header and footer). */
-function frameBodyRect(columns: number, rows: number): LayoutRect {
-  return {
-    row: Math.min(2, Math.max(0, rows)),
-    col: 0,
-    width: Math.max(0, columns),
-    height: Math.max(0, rows - 2),
-  };
-}
-
-
-/** Clip or pad a single line to exactly `width` visible columns. */
-function fitLine(line: string, width: number): string {
-  const clipped = clipToWidth(line, Math.max(0, width));
-  return clipped + ' '.repeat(Math.max(0, width - visibleLength(clipped)));
-}
-
-
-/** Check if a layout node (or its first descendant pane) is minimized. */
-function isPaneMinimized(node: FrameLayoutNode, visibility: PanelVisibilityState): boolean {
-  if (node.kind === 'pane') return isMinimized(visibility, node.paneId);
-  // For containers, check if all descendant panes are minimized
-  const paneIds = collectPaneIds(node);
-  return paneIds.length > 0 && paneIds.every((id) => isMinimized(visibility, id));
-}
-
-/** Merge two read-only maps into a new mutable map. */
-function mergeMaps<K, V>(a: ReadonlyMap<K, V>, b: ReadonlyMap<K, V>): Map<K, V> {
-  const out = new Map<K, V>();
-  for (const [k, v] of a) out.set(k, v);
-  for (const [k, v] of b) out.set(k, v);
-  return out;
-}
-
-/** Translate a layout rect by the given row and column offsets. */
-function offsetRect(rect: LayoutRect, rowOffset: number, colOffset: number): LayoutRect {
-  return {
-    row: rowOffset + rect.row,
-    col: colOffset + rect.col,
-    width: rect.width,
-    height: rect.height,
-  };
-}
-
-/** Convert a binding's key combo into a synthetic KeyMsg for dispatch. */
-function comboToMsg(binding: BindingInfo): KeyMsg {
-  return {
-    type: 'key',
-    key: binding.combo.key,
-    ctrl: binding.combo.ctrl,
-    alt: binding.combo.alt,
-    shift: binding.combo.shift,
-  };
-}
-
-/** Type guard: is this message a frame-internal action wrapper? */
-function isFrameScopedMsg(value: unknown): value is FrameScopedMsg {
-  return typeof value === 'object'
-    && value !== null
-    && FRAME_MSG_TOKEN in value
-    && (value as FrameScopedMsg)[FRAME_MSG_TOKEN] === true;
-}
-
-/** Wrap a frame action into a FrameScopedMsg for the update loop. */
-function wrapFrameMsg(action: FrameAction): FrameScopedMsg {
-  return {
-    [FRAME_MSG_TOKEN]: true,
-    action,
-  };
-}
-
-/** Create a command that immediately resolves with the given message. */
-function emitMsg<Msg>(msg: Msg): Cmd<Msg> {
-  return () => Promise.resolve(msg);
-}
-
-/** Create a command that emits a page-scoped message. */
-function emitMsgForPage<Msg>(pageId: string, msg: Msg): Cmd<Msg> {
-  return async () => wrapPageMsg(pageId, msg);
-}
-
-/** Wrap a page-level command so its emitted messages are tagged with the page ID. */
-function wrapCmdForPage<Msg>(pageId: string, cmd: Cmd<Msg>): Cmd<Msg> {
-  return async (emit) => {
-    const result = await cmd((msg) => emit(wrapPageMsg(pageId, msg) as unknown as Msg));
-    if (result === undefined || result === QUIT) return result;
-    return wrapPageMsg(pageId, result as Msg);
-  };
-}
-
-/** Tag a user message with its originating page ID. */
-function wrapPageMsg<Msg>(pageId: string, msg: Msg): Msg {
-  return {
-    [PAGE_MSG_TOKEN]: true,
-    pageId,
-    msg,
-  } as unknown as Msg;
-}
-
-/** Type guard: is this message a page-scoped wrapper? */
-function isPageScopedMsg<Msg>(value: unknown): value is PageScopedMsg<Msg> {
-  return typeof value === 'object'
-    && value !== null
-    && PAGE_MSG_TOKEN in value
-    && (value as PageScopedMsg<Msg>)[PAGE_MSG_TOKEN] === true;
-}
-
-/**
- * Create a TEA command that drives transition re-renders via `setInterval`.
- *
- * Each tick emits a frame-scoped 'transition' message. The actual progress
- * is computed from wall-clock time in the update handler, not from the
- * interval count. The interval just schedules re-renders.
- */
-function createTransitionTickCmd<Msg>(durationMs: number, generation: number): Cmd<Msg> {
-  return (emit) =>
-    new Promise<void>((resolve) => {
-      if (durationMs <= 0) {
-        emit(wrapFrameMsg({ type: 'transition-complete', generation } as FrameAction) as unknown as Msg);
-        resolve();
-        return;
-      }
-
-      const startMs = Date.now();
-      const intervalMs = Math.round(1000 / 60);
-
-      const id = setInterval(() => {
-        const elapsed = Date.now() - startMs;
-        const rawProgress = Math.min(1, elapsed / durationMs);
-
-        emit(wrapFrameMsg({ type: 'transition', progress: rawProgress, generation } as FrameAction) as unknown as Msg);
-
-        if (rawProgress >= 1) {
-          clearInterval(id);
-          emit(wrapFrameMsg({ type: 'transition-complete', generation } as FrameAction) as unknown as Msg);
-          resolve();
-        }
-      }, intervalMs);
-    });
-}
-
-/** Render a page's layout tree within the frame body rect. */
-function renderPageContent<PageModel, Msg>(
-  pageId: string,
-  model: InternalFrameModel<PageModel, Msg>,
-  bodyRect: LayoutRect,
-  pagesById: Map<string, FramePage<PageModel, Msg>>,
-): RenderResult {
-  const page = pagesById.get(pageId)!;
-  const pageModel = model.pageModels[pageId]!;
-  const renderCtx: RenderContext<PageModel, Msg> = {
-    model,
-    pageId,
-    focusedPaneId: model.focusedPaneByPage[pageId],
-    scrollByPane: model.scrollByPage[pageId] ?? {},
-    visibility: model.minimizedByPage[pageId] ?? createPanelVisibilityState(),
-    dockState: model.dockStateByPage[pageId] ?? createPanelDockState(),
-  };
-  return renderFrameNode(page.layout(pageModel), bodyRect, renderCtx);
-}
-
-/**
- * Split a styled multiline string into a 2D grid of single-column characters.
- * Each cell is a fully-styled string (including resets).
- */
-function stringToGrid(str: string, width: number, height: number): string[][] {
-  const lines = str.split('\n');
-  const grid: string[][] = [];
-
-  for (let y = 0; y < height; y++) {
-    const line = lines[y] ?? '';
-    grid.push(tokenizeAnsi(line, width));
-  }
-  return grid;
-}
-
-/** Apply a transition shader to blend between the previous and next page views. */
-function renderTransition(
-  prev: string,
-  next: string,
-  style: PageTransition,
-  progress: number,
-  width: number,
-  height: number,
-  ctx: BijouContext,
-): string {
-  const shader = typeof style === 'function' ? style : TRANSITION_SHADERS[style];
-  if (!shader) return next;
-
-  const prevGrid = stringToGrid(prev, width, height);
-  const nextGrid = stringToGrid(next, width, height);
-  const lines: string[] = [];
-
-  for (let y = 0; y < height; y++) {
-    let line = '';
-    for (let x = 0; x < width; x++) {
-      // Shared stable-ish pseudo-random seed based on coordinates
-      const seed = Math.sin(x * 12.9898 + y * 78.233) * 43758.5453;
-      const rand = seed - Math.floor(seed);
-
-      const result = shader({ x, y, width, height, progress, rand, ctx });
-      const showNext = result.showNext;
-      const charOverride = result.char;
-
-      if (charOverride !== undefined) {
-        line += charOverride;
-      } else {
-        line += (showNext ? nextGrid[y]?.[x] : prevGrid[y]?.[x]) ?? ' ';
-      }
-    }
-    lines.push(line);
-  }
-
-  return lines.join('\n');
 }
